Use 'waitfor' if it is valid when checking if we should call the upload function.
When uploading to an sftp server, after writing we have to read from the
socket before the write will complete. During this time we don't want to
write, so libssh2 sets the select to read (not write). Without this
change we never get into the 'readwrite_upload' block, as selectres will
never be marked for writing.